End of Section 18(Unit of work pattern)

+++ client/src/app/messages/messages.component.ts
@@ -3,6 +3,7 @@ import { Message } from '../_models/message';
 import { Pagination } from '../_models/Pagination';
 import { MessageService } from '../_services/message.service';
 import { DomSanitizer } from '@angular/platform-browser';
+import { ConfirmService } from '../_services/confirm.service';
 
 @Component({
   selector: 'app-messages',
@@ -17,7 +18,8 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   isLoading = false;
 
-  constructor(private messageService: MessageService, private sanitizer: DomSanitizer) { }
+  constructor(private messageService: MessageService, private sanitizer: DomSanitizer,
+    private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -34,9 +36,14 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m => m.id == id), 1)
-    })
+    this.confirmService.confirm("Confirm Delete message", "This cannot be reverted")
+      .subscribe(result => {
+        if (result) {
+          this.messageService.deleteMessage(id).subscribe(() => {
+            this.messages.splice(this.messages.findIndex(m => m.id == id), 1)
+          });
+        }
+      });
   }
 
   pageChanged(event: any) {
